fix moving a child item (inside tree)
inventory, staff, courses and subjects use a
reflexive association (parent(s)/child(ren)) to form tree
struktures.

there was a bug when changing an item's "parent", ie. to
move a staff from one organisation to another. it did not
manifest in db records, but journal records and UI tree
views (when moving, the item appears twice).