disable "blink" timer
we had showcase eCRFs that demonstrated fancy blinking editcheck messages
for demonstration purposes. the support for the "blink" attribute was dropped
in browser quite some time ago and therefore a workaround by toggling visibility
with a timer was placed.

removing this now, since it is not in use in the regular UI.