prevent "hanging" login.jsf
in situations with 2+ login prompts open, the user can log in with one, and
returns to the page prior to session expiry as desired.

when pressing the login buttin in a remaining login pompts, the it however did
not forward but remains. this was confusing user for long time.

it now forwards in this case as well. the provided credentials are not
authenticated, can be even empty.