re-encrypter helper objects - ctsms.security.reencrypt package
encrypted data of an entity can be categorized:

- field (serialized primitive value): iv + ciphertext + hash
- data (larger byte[] field or field w/o hash): iv + ciphertext
- file: iv + external file (on disk)

each category above is covered by a a reencrypter base class to derive
from anonymously (no lambda). the instances are stateless and
encapsulate decryption of the value with an (old) department's SecretKey
and subsequent encryption with a (new) department's SecretKey.

an entity's dao will have a list of 0 or more re-encrypter instances,
which is iterated by (a dao's) reEncrypt() function.