option to disable particular entity classes from VO caching
when handling with collections of evicted entities (showing non-unique
id's e. as a result of left joins), the VO caching would jump in and
re-use the same VO for a given entity id.

this must not happen eg. when converting a collection of visit schedule
appointments, otherwise they all would show the same calculated dates.