show "Cert." column in training record PDF only if needed
in the Traning Record PDF, tables showing lists of participated courses
have a "Cert." column, to indicate courses that are expected to have a
written Certificate document.

in order to save up horizontal table canvas, that "Cert." column should
be rendered only if the list will contain at least one course with a
expected cert.